clarify cron scheduler targets
