+++ gateway/src/fs/index.ts
@@ -5,6 +5,7 @@
 export { GsvFs } from "./gsv-fs";
 export type { ExtendedStat } from "./gsv-fs";
 export type { KernelRefs } from "./refs";
+export { requestProcessView } from "./refs";
 export type {
   MountBackend,
   ExtendedMountStat,
