+++ gateway/src/fs/refs.ts
@@ -5,6 +5,8 @@ import type { CronFileService } from "../kernel/crontab";
 import type { DeviceRegistry } from "../kernel/devices";
 import type { ProcessRegistry } from "../kernel/processes";
 import type { PackageStore } from "../kernel/packages";
+import type { RequestFrame } from "../protocol/frames";
+import { sendFrameToProcess } from "../shared/utils";
 import type { ArgsOf, ResultOf } from "../syscalls";
 import type { ScheduleRecord, ScheduleRunHistoryEntry } from "../syscalls/scheduler";
 
@@ -23,6 +25,27 @@ export type ProcessViewRequest = <S extends ProcessViewCall>(
   call: S,
   args: ArgsOf<S>,
 ) => Promise<ResultOf<S>>;
+
+export async function requestProcessView<S extends ProcessViewCall>(
+  pid: string,
+  call: S,
+  args: ArgsOf<S>,
+): Promise<ResultOf<S>> {
+  const frame = {
+    type: "req",
+    id: crypto.randomUUID(),
+    call,
+    args,
+  } as RequestFrame;
+  const response = await sendFrameToProcess(pid, frame);
+  if (!response || response.type !== "res") {
+    throw new Error(`${call} did not return a response`);
+  }
+  if (!response.ok) {
+    throw new Error(response.error.message);
+  }
+  return response.data as ResultOf<S>;
+}
 
 export type ScheduleViewStore = {
   list(args: {
